+++ App.tsx
@@ -1,19 +1,3 @@
-import React from "react";
-import { Provider } from "react-redux";
-import { applyMiddleware, createStore } from "redux";
-import createSagaMiddleware from "redux-saga";
-import AppStateChangedContainer from "./src/AppStateChangedContainer";
-import combinedReducers from "./src/redux/reducers";
-import rootSaga from "./src/redux/sagas";
-
-const sagaMiddleware = createSagaMiddleware();
-const store = createStore(combinedReducers, applyMiddleware(sagaMiddleware));
-sagaMiddleware.run(rootSaga);
-
-const App = () => (
-    <Provider store={store}>
-        <AppStateChangedContainer />
-    </Provider>
-);
+import App from "./src/App";
 
 export default App;
